Start a bit of refactoring/consolidation in prep for using Arrow IO interfaces
