+++ cpp/src/parquet/bloom_filter.h
@@ -27,7 +27,7 @@
 #include "parquet/hasher.h"
 #include "parquet/types.h"
 #include "parquet/util/memory.h"
-#include "parquet/util/visibility.h"
+#include "parquet/platform.h"
 
 namespace arrow {
