+++ cpp/src/parquet/column_reader.h
@@ -33,7 +33,7 @@
 #include "parquet/schema.h"
 #include "parquet/types.h"
 #include "parquet/util/memory.h"
-#include "parquet/util/visibility.h"
+#include "parquet/platform.h"
 
 namespace arrow {
