+++ cpp/src/parquet/column_writer.h
@@ -29,7 +29,7 @@
 #include "parquet/schema.h"
 #include "parquet/types.h"
 #include "parquet/util/memory.h"
-#include "parquet/util/visibility.h"
+#include "parquet/platform.h"
 
 namespace arrow {
