+++ cpp/src/parquet/encoding.h
@@ -30,7 +30,7 @@
 #include "parquet/exception.h"
 #include "parquet/types.h"
 #include "parquet/util/memory.h"
-#include "parquet/util/visibility.h"
+#include "parquet/platform.h"
 
 namespace parquet {
