+++ cpp/src/parquet/file_writer.h
@@ -28,7 +28,7 @@
 #include "parquet/metadata.h"
 #include "parquet/properties.h"
 #include "parquet/schema.h"
-#include "parquet/util/visibility.h"
+#include "parquet/platform.h"
 
 namespace arrow {
