+++ cpp/src/parquet/metadata.h
@@ -28,7 +28,7 @@
 
 #include "parquet/properties.h"
 #include "parquet/types.h"
-#include "parquet/util/visibility.h"
+#include "parquet/platform.h"
 
 namespace parquet {
