+++ cpp/src/parquet/statistics.h
@@ -22,11 +22,10 @@
 #include <memory>
 #include <string>
 
+#include "parquet/platform.h"
 #include "parquet/schema.h"
 #include "parquet/types.h"
-#include "parquet/util/macros.h"
 #include "parquet/util/memory.h"
-#include "parquet/util/visibility.h"
 
 namespace parquet {
