+++ cpp/src/parquet/types.cc
@@ -25,6 +25,7 @@
 #include <utility>
 
 #include "arrow/util/checked_cast.h"
+#include "arrow/util/compression.h"
 #include "arrow/util/logging.h"
 
 #include "parquet/exception.h"
@@ -32,8 +33,38 @@
 #include "parquet/types.h"
 
 using ::arrow::internal::checked_cast;
+using arrow::util::Codec;
 
 namespace parquet {
+
+std::unique_ptr<Codec> GetCodecFromArrow(Compression::type codec) {
+  std::unique_ptr<Codec> result;
+  switch (codec) {
+    case Compression::UNCOMPRESSED:
+      break;
+    case Compression::SNAPPY:
+      PARQUET_THROW_NOT_OK(Codec::Create(::arrow::Compression::SNAPPY, &result));
+      break;
+    case Compression::GZIP:
+      PARQUET_THROW_NOT_OK(Codec::Create(::arrow::Compression::GZIP, &result));
+      break;
+    case Compression::LZO:
+      PARQUET_THROW_NOT_OK(Codec::Create(::arrow::Compression::LZO, &result));
+      break;
+    case Compression::BROTLI:
+      PARQUET_THROW_NOT_OK(Codec::Create(::arrow::Compression::BROTLI, &result));
+      break;
+    case Compression::LZ4:
+      PARQUET_THROW_NOT_OK(Codec::Create(::arrow::Compression::LZ4, &result));
+      break;
+    case Compression::ZSTD:
+      PARQUET_THROW_NOT_OK(Codec::Create(::arrow::Compression::ZSTD, &result));
+      break;
+    default:
+      break;
+  }
+  return result;
+}
 
 std::string FormatStatValue(Type::type parquet_type, const std::string& val) {
   std::stringstream result;
